requirements.txt
# Core Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary
python-dotenv

# LLM & AI
httpx>=0.27.0
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Auth
bcrypt>=5.0.0

# Structured Logging
structlog

# Data Validation
pydantic>=2.0.0

# Testing (dev only)
pytest
pytest-asyncio
pytest-cov